feat(GIST-78): added ABAC on gists

+++ gists/controller.go
@@ -5,7 +5,9 @@ import (
 	"github.com/gofiber/fiber/v2/log"
 )
 
-type GistControllerImpl struct{}
+type GistControllerImpl struct {
+	gist_guard GistGuard
+}
 
 type GistSaveValidator struct {
 	Name        string `json:"name"`
@@ -13,6 +15,7 @@ type GistSaveValidator struct {
 	OrgID       string `json:"org_id,omitempty"`
 	Language    string `json:"language,omitempty"`
 	Description string `json:"description,omitempty"`
+	Visibility  string `json:"visibility,omitempty"`
 }
 
 func (g *GistControllerImpl) Save() fiber.Handler {
@@ -23,7 +26,13 @@ func (g *GistControllerImpl) Save() fiber.Handler {
 		if err := c.BodyParser(g); err != nil {
 			return c.Status(400).SendString("Request must be valid JSON with fields name and content as text")
 		}
-		gist, err := GistService.Save(g.Name, g.Content, owner_id, g.OrgID, g.Language, g.Description)
+
+		visibility := "public"
+		if g.Visibility == "private" {
+			visibility = "private"
+		}
+
+		gist, err := GistService.Save(g.Name, g.Content, owner_id, g.OrgID, g.Language, g.Description, visibility)
 		if err != nil {
 			return c.Status(500).SendString(err.Error())
 		}
@@ -34,13 +43,24 @@ func (g *GistControllerImpl) Save() fiber.Handler {
 
 func (g *GistControllerImpl) UpdateName() fiber.Handler {
 	return func(c *fiber.Ctx) error {
-		g := new(GistSaveValidator)
-		if err := c.BodyParser(g); err != nil {
+		g_body := new(GistSaveValidator)
+		if err := c.BodyParser(g_body); err != nil {
 			return c.Status(400).SendString("Request must be valid JSON with fields name and content as text")
 		}
 
 		owner_id := c.Locals("pub").(string)
-		gist, err := GistService.UpdateName(c.Params("id"), g.Name, owner_id)
+
+		can_edit, err := g.gist_guard.CanEdit(c.Params("id"), owner_id)
+
+		if err != nil {
+			return c.Status(500).SendString(err.Error())
+		}
+
+		if !can_edit {
+			return c.Status(403).SendString("You do not have permission to edit this gist")
+		}
+
+		gist, err := GistService.UpdateName(c.Params("id"), g_body.Name, owner_id)
 		if err != nil {
 			if err == ErrGistNotFound {
 				return c.Status(404).SendString(err.Error())
@@ -67,6 +87,16 @@ func (g *GistControllerImpl) FindByID() fiber.Handler {
 	return func(c *fiber.Ctx) error {
 		owner_id := c.Locals("pub").(string)
 
+		can_read, err := g.gist_guard.CanRead(c.Params("id"), owner_id)
+
+		if err != nil {
+			return c.Status(500).SendString(err.Error())
+		}
+
+		if !can_read {
+			return c.Status(403).SendString("You do not have permission to read this gist")
+		}
+
 		gist, err := GistService.FindByID(c.Params("id"), owner_id)
 		if err != nil {
 			return c.Status(404).SendString(err.Error())
@@ -85,6 +115,16 @@ func (g *GistControllerImpl) RawFindByID() fiber.Handler {
 	return func(c *fiber.Ctx) error {
 		owner_id := c.Locals("pub").(string)
 
+		can_read, err := g.gist_guard.CanRead(c.Params("id"), owner_id)
+
+		if err != nil {
+			return c.Status(500).SendString(err.Error())
+		}
+
+		if !can_read {
+			return c.Status(403).SendString("You do not have permission to read this gist")
+		}
+
 		gist, err := GistService.FindByID(c.Params("id"), owner_id)
 		if err != nil {
 			return c.Status(404).SendString(err.Error())
@@ -96,13 +136,24 @@ func (g *GistControllerImpl) RawFindByID() fiber.Handler {
 
 func (g *GistControllerImpl) UpdateContent() fiber.Handler {
 	return func(c *fiber.Ctx) error {
-		g := new(GistSaveValidator)
-		if err := c.BodyParser(g); err != nil {
+		g_body := new(GistSaveValidator)
+		if err := c.BodyParser(g_body); err != nil {
 			return c.Status(400).SendString("Request must be valid JSON with fields name and content as text")
 		}
 
 		owner_id := c.Locals("pub").(string)
-		gist, err := GistService.UpdateContent(c.Params("id"), g.Content, owner_id)
+
+		can_edit, err := g.gist_guard.CanEdit(c.Params("id"), owner_id)
+
+		if err != nil {
+			return c.Status(500).SendString(err.Error())
+		}
+
+		if !can_edit {
+			return c.Status(403).SendString("You do not have permission to edit this gist")
+		}
+
+		gist, err := GistService.UpdateContent(c.Params("id"), g_body.Content, owner_id)
 		if err != nil {
 			if err == ErrGistNotFound {
 				return c.Status(404).SendString(err.Error())
@@ -115,12 +166,23 @@ func (g *GistControllerImpl) UpdateContent() fiber.Handler {
 
 func (g *GistControllerImpl) UpdateLanguage() fiber.Handler {
 	return func(c *fiber.Ctx) error {
-		g := new(GistSaveValidator)
-		if err := c.BodyParser(g); err != nil {
+		g_body := new(GistSaveValidator)
+		if err := c.BodyParser(g_body); err != nil {
 			return c.Status(400).SendString("Request must be valid JSON with fields name and content as text")
 		}
 		owner_id := c.Locals("pub").(string)
-		gist, err := GistService.UpdateLanguage(c.Params("id"), g.Language, owner_id)
+
+		can_edit, err := g.gist_guard.CanEdit(c.Params("id"), owner_id)
+
+		if err != nil {
+			return c.Status(500).SendString(err.Error())
+		}
+
+		if !can_edit {
+			return c.Status(403).SendString("You do not have permission to edit this gist")
+		}
+
+		gist, err := GistService.UpdateLanguage(c.Params("id"), g_body.Language, owner_id)
 		if err != nil {
 			if err == ErrGistNotFound {
 				return c.Status(404).SendString(err.Error())
@@ -133,12 +195,23 @@ func (g *GistControllerImpl) UpdateLanguage() fiber.Handler {
 
 func (g *GistControllerImpl) UpdateDescription() fiber.Handler {
 	return func(c *fiber.Ctx) error {
-		g := new(GistSaveValidator)
-		if err := c.BodyParser(g); err != nil {
+		g_body := new(GistSaveValidator)
+		if err := c.BodyParser(g_body); err != nil {
 			return c.Status(400).SendString("Request must be valid JSON with fields name and content as text")
 		}
 		owner_id := c.Locals("pub").(string)
-		gist, err := GistService.UpdateDescription(c.Params("id"), g.Description, owner_id)
+
+		can_edit, err := g.gist_guard.CanEdit(c.Params("id"), owner_id)
+
+		if err != nil {
+			return c.Status(500).SendString(err.Error())
+		}
+
+		if !can_edit {
+			return c.Status(403).SendString("You do not have permission to edit this gist")
+		}
+
+		gist, err := GistService.UpdateDescription(c.Params("id"), g_body.Description, owner_id)
 		if err != nil {
 			if err == ErrGistNotFound {
 				return c.Status(404).SendString(err.Error())
@@ -152,6 +225,17 @@ func (g *GistControllerImpl) UpdateDescription() fiber.Handler {
 func (g *GistControllerImpl) Delete() fiber.Handler {
 	return func(c *fiber.Ctx) error {
 		owner_id := c.Locals("pub").(string)
+
+		can_edit, err := g.gist_guard.CanEdit(c.Params("id"), owner_id)
+
+		if err != nil {
+			return c.Status(500).SendString(err.Error())
+		}
+
+		if !can_edit {
+			return c.Status(403).SendString("You do not have permission to edit this gist")
+		}
+
 		if err := GistService.Delete(c.Params("id"), owner_id); err != nil {
 			if err == ErrGistNotFound {
 				return c.Status(404).SendString(err.Error())
